typo

+++ tracking/main.py
@@ -3,7 +3,7 @@ import random
 import vessl
 
 epochs = int(os.environ['epochs'])
-lr = int(os.environ['lr'])
+lr = float(os.environ['lr'])
 offset = random.random() / 5
 
 for epoch in range(2, epochs):
